backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg[binary]==3.2.1
python-dotenv==1.0.1
scikit-learn==1.5.2
joblib==1.4.2
pandas